refactored disposing containers
