+++ sdk/storage/Azure.Storage.Blobs/tests/TenantDiscoveryBlobBaseClientTests.cs
@@ -4,9 +4,9 @@
 using System.Threading.Tasks;
 using Azure.Core.TestFramework;
 using Azure.Storage.Blobs.Specialized;
+using Azure.Storage.Blobs.Tests;
 using Azure.Storage.Test.Shared;
 using NUnit.Framework;
-using static Azure.Storage.Blobs.Tests.ClientBuilderExtensions;
 
 namespace Azure.Storage.Blobs.Tests
 {
