+++ sdk/storage/Azure.Storage.Files.DataLake/tests/ClientBuilderExtensions.cs
@@ -77,31 +77,5 @@ namespace Azure.Storage.Files.DataLake.Tests
 
             return new DisposingFileSystem(fileSystem);
         }
-
-        public class DisposingFileSystem : IAsyncDisposable
-        {
-            public DataLakeFileSystemClient FileSystem;
-
-            public DisposingFileSystem(DataLakeFileSystemClient fileSystem)
-            {
-                FileSystem = fileSystem;
-            }
-
-            public async ValueTask DisposeAsync()
-            {
-                if (FileSystem != null)
-                {
-                    try
-                    {
-                        await FileSystem.DeleteIfExistsAsync();
-                        FileSystem = null;
-                    }
-                    catch
-                    {
-                        // swallow the exception to avoid hiding another test failure
-                    }
-                }
-            }
-        }
     }
 }
