+++ sdk/storage/Azure.Storage.Files.DataLake/tests/FileSystemClientTests.cs
@@ -14,7 +14,6 @@ using Azure.Storage.Sas;
 using Azure.Storage.Test;
 using Moq;
 using NUnit.Framework;
-using static Azure.Storage.Files.DataLake.Tests.ClientBuilderExtensions;
 
 namespace Azure.Storage.Files.DataLake.Tests
 {
