+++ sdk/storage/Azure.Storage.Files.Shares/tests/CopySourceAuthTests.cs
@@ -13,7 +13,6 @@ using Azure.Storage.Files.Shares.Models;
 using Azure.Storage.Test;
 using Azure.Storage.Test.Shared;
 using NUnit.Framework;
-using static Azure.Storage.Files.Shares.Tests.ClientBuilderExtensions;
 
 namespace Azure.Storage.Files.Shares.Tests
 {
