+++ sdk/storage/Azure.Storage.Files.Shares/tests/FileSasBuilderTests.cs
@@ -7,7 +7,6 @@ using Azure.Core.TestFramework;
 using Azure.Storage.Sas;
 using Azure.Storage.Test;
 using NUnit.Framework;
-using static Azure.Storage.Files.Shares.Tests.ClientBuilderExtensions;
 using TestConstants = Azure.Storage.Test.TestConstants;
 
 namespace Azure.Storage.Files.Shares.Tests
