+++ sdk/storage/Azure.Storage.Files.Shares/tests/ServiceClientTests.cs
@@ -11,7 +11,6 @@ using Azure.Storage.Sas;
 using Azure.Storage.Test;
 using Moq;
 using NUnit.Framework;
-using static Azure.Storage.Files.Shares.Tests.ClientBuilderExtensions;
 
 namespace Azure.Storage.Files.Shares.Tests
 {
