+++ sdk/storage/Azure.Storage.Queues/tests/ClientSideEncryptionTests.cs
@@ -21,7 +21,6 @@ using Azure.Storage.Queues.Tests;
 using Azure.Storage.Test;
 using Moq;
 using NUnit.Framework;
-using static Azure.Storage.Queues.Tests.ClientBuilderExtensions;
 using static Moq.It;
 
 namespace Azure.Storage.Queues.Test
