+++ sdk/storage/Azure.Storage.Queues/tests/QueueSasBuilderTests.cs
@@ -8,7 +8,6 @@ using Azure.Storage.Queues.Tests;
 using Azure.Storage.Sas;
 using Azure.Storage.Test;
 using NUnit.Framework;
-using static Azure.Storage.Queues.Tests.ClientBuilderExtensions;
 using TestConstants = Azure.Storage.Test.TestConstants;
 
 namespace Azure.Storage.Queues.Test
