+++ sdk/storage/Azure.Storage.Queues/tests/ServiceClientTests.cs
@@ -13,7 +13,6 @@ using Azure.Storage.Sas;
 using Azure.Storage.Test;
 using Moq;
 using NUnit.Framework;
-using static Azure.Storage.Queues.Tests.ClientBuilderExtensions;
 
 namespace Azure.Storage.Queues.Test
 {
